requirements.txt
certifi==2025.1.31
influxdb3-python==0.12.0
numpy==2.2.4
pandas==2.2.3
pyarrow==19.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
reactivex==4.0.4
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
